Code review clean ups:
1. Use if statements rather than Option#ofNullable.
2. A few Javadoc tweaks.
3. Move new test to framework-test.
