+++ framework-test/src/main/java/org/checkerframework/framework/test/TestRootDirectory.java
@@ -6,7 +6,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** Defines the root directory path to search for tests java sources under. */
+/** Defines the path to the directory which holds test java files. */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
