+++ framework-test/src/test/java/org/checkerframework/framework/test/test/junit/AlternateTestRootPerFileWithFilesTest.java
@@ -1,4 +1,4 @@
-package org.checkerframework.framework.test.junit;
+package org.checkerframework.framework.test.test.junit;
 
 import java.io.File;
 import java.util.List;
