+++ framework-test/tests-alt/alt-dir-b/Issue6125B.java
